Null out `generateSnapshot` function after snapshot has been generated
This fixes the slowness we were observing in Atom when opening the
DevTools and trying to use the Console autocomplete feature.

By deleting `generateSnapshot` from the global scope, we will prevent it
from being serialized every time the DevTools process fetches
autocomplete suggestions from the renderer process. Serializing it can
be expensive because that function contains the entire source code of
the bundle, and that gets sent as a string via IPC to the DevTools
process in order to provide the function preview.